change filters to ts
