+++ src/components/FilterModals/DeptFilterModal.tsx
@@ -14,18 +14,26 @@ import {
   useBreakpointValue,
   useColorModeValue,
 } from "@chakra-ui/react";
-import { type_list, code_map } from "data/course_type";
-import FilterElement from "components/FilterModals/components/FilterElement";
 import React, { useMemo } from "react";
+import { college_map } from "data/college";
+import { dept_list_bachelor_only } from "data/department";
+import FilterElement from "components/FilterModals/components/FilterElement";
 import { useCourseSearchingContext } from "components/Providers/CourseSearchingProvider";
 import { reportEvent } from "utils/ga";
 
-function CategoryFilterModal({
+export interface DeptFilterModalProps {
+  readonly title: string;
+  readonly isEnabled: boolean;
+  readonly selectedDept: string[];
+  readonly setSelectedDept: (time: string[]) => void;
+}
+
+function DeptFilterModal({
   title,
   isEnabled,
-  selectedType,
-  setSelectedType,
-}) {
+  selectedDept,
+  setSelectedDept,
+}: DeptFilterModalProps) {
   const headingColor = useColorModeValue("heading.light", "heading.dark");
   const { searchFilters, setSearchFilters } = useCourseSearchingContext();
   const { isOpen, onOpen, onClose } = useDisclosure();
@@ -33,7 +41,7 @@ function CategoryFilterModal({
 
   const onOpenModal = () => {
     // overwrite local states by redux store
-    setSelectedType(searchFilters.category);
+    setSelectedDept(searchFilters.department);
     onOpen();
   };
 
@@ -41,33 +49,35 @@ function CategoryFilterModal({
     // fire when click "X" or outside of modal
     // overwrite local state by redux state
     onClose();
-    setSelectedType(searchFilters.category);
+    setSelectedDept(searchFilters.department);
   };
 
   const onSaveEditing = () => {
     // fire when click "Save"
     // overwrite redux state by local state
-    setSearchFilters({ ...searchFilters, category: selectedType });
+    setSearchFilters({ ...searchFilters, department: selectedDept });
     onClose();
   };
 
   const onResetEditing = () => {
     // fire when click "Reset"
     // set local state to empty array
-    setSelectedType([]);
+    setSelectedDept([]);
   };
 
   const modalBody = useMemo(
     () => (
-      <React.Fragment key={`modalBody`}>
-        {Object.keys(code_map).map((code_map_key, index) => {
-          const categories = type_list.filter(
-            (type) => type.code[0] === code_map_key
+      <>
+        {Object.keys(college_map).map((college_key, index) => {
+          const departments = dept_list_bachelor_only.filter(
+            (dept) => dept.code[0] === college_key
           );
+          if (departments.length === 0) {
+            return null;
+          }
           return (
-            <React.Fragment key={`${code_map_key}`}>
+            <React.Fragment key={`${college_key}-${index}`}>
               <Flex
-                key={`${code_map_key}-${index}`}
                 px="2"
                 h="40px"
                 flexDirection="column"
@@ -79,30 +89,30 @@ function CategoryFilterModal({
                 bg={modalBgColor}
               >
                 <Heading fontSize="2xl" color={headingColor}>
-                  {code_map[code_map_key].name}
+                  {college_key + " " + college_map[college_key].name}
                 </Heading>
                 <Divider />
               </Flex>
-              {categories
-                .filter((type) => !selectedType.includes(type.id))
-                .map((type, typeIdx) => (
+              {departments
+                .filter((dept) => !selectedDept.includes(dept.code))
+                .map((dept, dept_index) => (
                   <FilterElement
-                    key={`${type.id}-${typeIdx}`}
-                    id={type.code}
-                    name={type.full_name}
+                    key={`${dept.code}-${dept_index}-modalBody`}
+                    id={dept.code}
+                    name={dept.full_name}
                     selected={false}
                     onClick={() => {
-                      setSelectedType([...selectedType, type.id]);
-                      reportEvent("filter_category", "click", type.id);
+                      setSelectedDept([...selectedDept, dept.code]);
+                      reportEvent("filter_department", "click", dept.code);
                     }}
                   />
                 ))}
             </React.Fragment>
           );
         })}
-      </React.Fragment>
+      </>
     ),
-    [selectedType, setSelectedType, headingColor, modalBgColor]
+    [selectedDept, setSelectedDept, headingColor, modalBgColor]
   );
 
   return (
@@ -112,7 +122,7 @@ function CategoryFilterModal({
         isDisabled={!isEnabled}
         onClick={() => {
           onOpenModal();
-          reportEvent("filter_category", "click", "open_modal");
+          reportEvent("filter_department", "click", "open_modal");
         }}
       >
         {title}
@@ -142,7 +152,7 @@ function CategoryFilterModal({
                 mr={3}
                 onClick={() => {
                   onSaveEditing();
-                  reportEvent("filter_category", "click", "save_changes");
+                  reportEvent("filter_department", "click", "save_changes");
                 }}
               >
                 套用
@@ -152,7 +162,7 @@ function CategoryFilterModal({
                 variant="ghost"
                 onClick={() => {
                   onResetEditing();
-                  reportEvent("filter_category", "click", "reset_changes");
+                  reportEvent("filter_department", "click", "reset_changes");
                 }}
               >
                 重設
@@ -161,19 +171,19 @@ function CategoryFilterModal({
           </ModalHeader>
           <ModalCloseButton />
           <ModalBody overflow="auto" pt="0">
-            {type_list
-              .filter((type) => selectedType.includes(type.id))
-              .map((type, index) => (
+            {dept_list_bachelor_only
+              .filter((dept) => selectedDept.includes(dept.code))
+              .map((dept, index) => (
                 <FilterElement
-                  key={`${type.id}-${index}-modalHeader`}
-                  id={type.code}
-                  name={type.full_name}
+                  key={`${dept.code}-${index}-modalHeader`}
+                  id={dept.code}
+                  name={dept.full_name}
                   selected={true}
                   onClick={() => {
-                    setSelectedType(
-                      selectedType.filter((id) => id !== type.id)
+                    setSelectedDept(
+                      selectedDept.filter((code) => code !== dept.code)
                     );
-                    reportEvent("filter_category", "click", type.id);
+                    reportEvent("filter_department", "click", dept.code);
                   }}
                 />
               ))}
@@ -186,7 +196,7 @@ function CategoryFilterModal({
               mr={3}
               onClick={() => {
                 onSaveEditing();
-                reportEvent("filter_category", "click", "save_changes");
+                reportEvent("filter_department", "click", "save_changes");
               }}
             >
               套用
@@ -195,7 +205,7 @@ function CategoryFilterModal({
               variant="ghost"
               onClick={() => {
                 onResetEditing();
-                reportEvent("filter_category", "click", "reset_changes");
+                reportEvent("filter_department", "click", "reset_changes");
               }}
             >
               重設
@@ -207,4 +217,4 @@ function CategoryFilterModal({
   );
 }
 
-export default CategoryFilterModal;
+export default DeptFilterModal;
